GetCoursesAttendedByCurrentStudent metodu eklendi

+++ Core/OnionArch.Application/DTOs/CourseDTO.cs
@@ -1,8 +1,8 @@
 ﻿namespace OnionArch.Application.DTOs;
 public record CourseDTO
 {
-    public required long Id { get; init; }
-    public required string Title { get; set; }
-    public required string Description { get; set; }
-    public required decimal Price { get; set; }
+	public required long Id { get; init; }
+	public required string Title { get; set; }
+	public required string Description { get; set; }
+	public required double Price { get; set; }
 }
